+++ Headers/DebugServer2/Architecture/X86_64/CPUState.h
@@ -83,12 +83,12 @@ struct CPUState64 {
 
   struct {
     X87Register regs[8];
-    uint32_t fstw;
-    uint32_t fctw;
-    uint32_t ftag;
-    uint64_t firip;
-    uint64_t forip;
-    uint32_t fop;
+    uint16_t fstw;
+    uint16_t fctw;
+    uint8_t ftag;
+    uint32_t firip;
+    uint32_t forip;
+    uint16_t fop;
   } x87;
 
   union {
